+++ aiperf/common/service/base_service.py
@@ -16,17 +16,20 @@ import asyncio
 import logging
 import signal
 import uuid
+from abc import ABC
 
-from aiperf.common.comms.communication import BaseCommunication
+import setproctitle
+
+from aiperf.common.comms.base_communication import BaseCommunication
 from aiperf.common.comms.communication_factory import CommunicationFactory
 from aiperf.common.config.service_config import ServiceConfig
 from aiperf.common.enums import ServiceState
-from aiperf.common.models.messages import BaseMessage
-from aiperf.common.models.payloads import PayloadType
-from aiperf.common.service.abstract import AbstractBaseService
+from aiperf.common.models.message_models import BaseMessage
+from aiperf.common.models.payload_models import PayloadType
+from aiperf.common.service.abstract_base_service import AbstractBaseService
 
 
-class BaseService(AbstractBaseService):
+class BaseService(AbstractBaseService, ABC):
     """Base class for all AIPerf services, providing common functionality for
     communication, state management, and lifecycle operations.
 
@@ -54,6 +57,13 @@ class BaseService(AbstractBaseService):
 
         # Set to store signal handler tasks
         self._signal_tasks = set()
+
+        # noinspection PyBroadException
+        try:
+            setproctitle.setproctitle(f"aiperf {self.service_type} ({self.service_id})")
+        except:  # noqa: E722
+            # setproctitle is not available on all platforms, so we ignore the error
+            self.logger.debug("Failed to set process title, ignoring")
 
     @property
     def state(self) -> ServiceState:
